ML/requirements.txt
gitdb==4.0.5
GitPython==3.1.11
psycopg2==2.8.6
smmap==3.0.4
